+++ backend/src/main/java/com/erepnikov/service/mapper/UserMapper.java
@@ -49,7 +49,7 @@ public class UserMapper {
                 .collect(Collectors.toList());
     }
 
-    public User userFromId(Long id) {
+    public User userFromId(Integer id) {
         if (id == null) {
             return null;
         }
